third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-auto-flow-get-set-expected.txt

Test that setting and getting grid-auto-flow works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting auto-flow set through CSS
PASS window.getComputedStyle(gridAutoFlowNone, '').getPropertyValue('grid-auto-flow') is 'none'
PASS window.getComputedStyle(gridAutoFlowColumn, '').getPropertyValue('grid-auto-flow') is 'column'
PASS window.getComputedStyle(gridAutoFlowRow, '').getPropertyValue('grid-auto-flow') is 'row'
PASS window.getComputedStyle(gridAutoFlowColumns, '').getPropertyValue('grid-auto-flow') is 'none'
PASS window.getComputedStyle(gridAutoFlowRows, '').getPropertyValue('grid-auto-flow') is 'none'
PASS window.getComputedStyle(gridAutoFlowInherit, '').getPropertyValue('grid-auto-flow') is 'column'
PASS window.getComputedStyle(gridAutoFlowNoInherit, '').getPropertyValue('grid-auto-flow') is 'none'

Test the initial value
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'none'

Test getting and setting grid-auto-flow through JS
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'column'
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'row'

Test getting and setting bad values for grid-auto-flow through JS
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'row'

Test setting grid-auto-flow to 'initial' through JS
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'none'
PASS successfullyParsed is true

TEST COMPLETE
